Update since December 21

+++ EditTenant/run.ps1
@@ -4,36 +4,48 @@ using namespace System.Net
 param($Request, $TriggerMetadata)
 
 $APIName = $TriggerMetadata.FunctionName
-Log-Request -user $request.headers.'x-ms-client-principal' -API $APINAME  -message "Accessed this API" -Sev "Debug"
-$Results = [System.Collections.ArrayList]@()
-
-
-# Write to the Azure Functions log stream.
-Write-Host "PowerShell HTTP trigger function processed a request."
-
-# Interact with query parameters or the body of the request.
+Log-Request -user $request.headers.'x-ms-client-principal' -API $APINAME -message 'Accessed this API' -Sev 'Debug'
 
 $tenantDisplayName = $request.body.displayName
 $tenantDefaultDomainName = $request.body.defaultDomainName
 $Tenant = $request.body.tenantid
-$tenantObjID = $request.body.id
+$customerContextId = $request.body.customerId
 
+$tokens = try {
 
-$results = try {
-    $bodyToPatch = '{"displayName":"' + $tenantDisplayName + '","defaultDomainName":"' + $tenantDefaultDomainName + '",}'
-    $Request = New-GraphPOSTRequest -uri "https://graph.microsoft.com/beta/contracts/$tenantObjID" -type PATCH -body $bodyToPatch
-    Log-Request -user $request.headers.'x-ms-client-principal' -API $APINAME -tenant $($Tenant) -message "Edited tenant $($Tenant)" -Sev "Info"
-    Remove-CIPPCache
-
-    "Successfully amended details for $($Tenant) and cleared tenant cache<br>"
-
+    $AADGraphtoken = (Get-GraphToken -scope 'https://graph.windows.net/.default')
+    $allTenantsDetails = (Invoke-RestMethod -Method GET -Uri 'https://graph.windows.net/myorganization/contracts?api-version=1.6' -ContentType 'application/json' -Headers $AADGraphtoken)
+    $tenantObjectId = $allTenantsDetails.value | Where-Object { $_.customerContextId -eq $customerContextId } | Select-Object 'objectId'
 } catch {
-    "Failed to amend details for $($Tenant): $($_.ExceptionMessage) <br>"
-    Log-Request -user $request.headers.'x-ms-client-principal' -API $APINAME -tenant $($Tenant) -message "Failed amending details $($tenantDisplayName). Error: $($_.Exception.Message)" -Sev "Error"
-    continue
+    "Failed to retrieve list of tenants.  Error: $($_.ExceptionMessage)"
+    Log-Request -user $request.headers.'x-ms-client-principal' -API $APINAME -tenant $($tenantDisplayName) -message "Failed to retrieve list of tenants. Error: $($_.Exception.Message)" -Sev 'Error'
 }
 
-$body = [pscustomobject]@{"Results" = $results }
+
+$results = if ($tenantObjectId) {
+    try {
+        $bodyToPatch = '{"displayName":"' + $tenantDisplayName + '","defaultDomainName":"' + $tenantDefaultDomainName + '"}'
+        $patchTenant = (Invoke-RestMethod -Method PATCH -Uri "https://graph.windows.net/myorganization/contracts/$($tenantObjectId.objectId)?api-version=1.6" -Body $bodyToPatch -ContentType 'application/json' -Headers $AADGraphtoken)    
+        
+        if (Test-Path -Path '.\tenants.cache.json') {
+            try {
+                Remove-CIPPCache
+            }
+            catch { 
+                'There was an error removing the tenants cache for some reason.' 
+            } 
+        }
+        Log-Request -user $request.headers.'x-ms-client-principal' -API $APINAME -tenant $($Tenant) -message "Edited tenant $($Tenant)" -Sev 'Info'
+        "Successfully amended details for $($Tenant) and cleared tenant cache"
+    }
+    catch { 
+        "Failed to amend details for $($Tenant): $($_.ExceptionMessage)"
+        Log-Request -user $request.headers.'x-ms-client-principal' -API $APINAME -tenant $($Tenant) -message "Failed amending details $($Tenant). Error: $($_.Exception.Message)" -Sev 'Error'
+        continue
+    }
+}
+
+$body = [pscustomobject]@{'Results' = $results }
 
 # Associate values to output bindings by calling 'Push-OutputBinding'.
 Push-OutputBinding -Name Response -Value ([HttpResponseContext]@{
